Update minimal_loop paths and settings for /data/yayun

+++ signwriting_animation/scripts/minimal_loop.py
@@ -172,8 +172,8 @@ if __name__ == "__main__":
         fut  = batch["data"][:1].to(model.device)
         mask = cond["target_mask"][:1].to(model.device)
 
-        print("[CHECK] target_mask mean / sum:", mask.mean().item(), mask.sum().item())
-        
+        print("[CHECK] target_mask mean / sum:", mask.float().mean().item(), mask.float().sum().item())
+
         print("=== USING TEACHER-FORCED PATH ===")
         ts = torch.zeros(fut.size(0), dtype=torch.long, device=fut.device)
         in_seq = fut.tensor.clone() if hasattr(fut, "tensor") else fut.clone()
